backend/requirements.txt
requests
boto3
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic-settings
elasticsearch
celery
redis
python-multipart
torch
torchvision
torchaudio
openai-whisper
nltk==3.8.1
langdetect==1.0.9
psutil
